category detail screen added

+++ lib/ui/common/ui_helpers.dart
@@ -33,7 +33,7 @@ double screenWidth(BuildContext context) => MediaQuery.of(context).size.width;
 double screenHeight(BuildContext context) => MediaQuery.of(context).size.height;
 
 double screenHeightFraction(BuildContext context,
-        {int dividedBy = 1, double offsetBy = 0, double max = 3000}) =>
+        {double dividedBy = 1, double offsetBy = 0, double max = 3000}) =>
     min((screenHeight(context) - offsetBy) / dividedBy, max);
 
 double screenWidthFraction(BuildContext context,
